+++ kamon-executors/src/main/resources/reference.conf
@@ -6,7 +6,7 @@ kamon.executors {
 
 kanela {
   show-banner = true
-  debug-mode = false
+  debug-mode = true
   log-level = "INFO"
   modules {
     executors-module {
@@ -16,7 +16,10 @@ kanela {
         "kamon.executors.instrumentation.ExecutorInstrumentation"
       ]
       within = [
+        "com.google.common.util.concurrent..*",
         "java.util.concurrent..*",
+        "scala.concurrent..*",
+        "akka.dispatch..*",
         "kamon.executors..*"
       ]
     }
