+++ kamon-executors/src/main/scala/kamon/executors/util/ContextAwareExecutorService.scala
@@ -81,8 +81,8 @@ class ContextAwareExecutorService(underlying: ExecutorService) extends ExecutorS
   }
 }
 
-class ContextAwareRunnable(r: Runnable) extends Runnable with HasContext {
-  override val context: Context = Kamon.currentContext()
+class ContextAwareRunnable(r: Runnable) extends Runnable {
+  val context: Context = Kamon.currentContext()
 
   override def run(): Unit = {
     Kamon.withContext(context) {
@@ -91,8 +91,8 @@ class ContextAwareRunnable(r: Runnable) extends Runnable with HasContext {
   }
 }
 
-class ContextAwareCallable[A](c: Callable[A]) extends HasContext with Callable[A] {
-  override val context: Context = Kamon.currentContext()
+class ContextAwareCallable[A](c: Callable[A]) extends Callable[A] {
+  val context: Context = Kamon.currentContext()
 
   override def call(): A = {
     Kamon.withContext(context) {
