+++ kamon-executors/src/test/resources/application.conf
