+++ kamon-executors/src/test/scala/kamon/executors/ExecutorMetricsSpec.scala
