+++ kamon-executors/src/test/scala/kamon/executors/ExecutorsRegistrationSpec.scala
